third_party/blink/web_tests/platform/generic/http/tests/security/xss-DENIED-window-name-alert-expected.txt

CONSOLE MESSAGE: SecurityError: Blocked a frame with origin "null" from accessing a cross-origin frame.
